[FAB-7241] Fix chaincode-devmode
This CR replaces the expired certificates with the msp library
in fabric/sampleconfig.

Regenerate the orderer.block and myc.tx artifact using the
configtx.yaml in fabric/sampleconfig.  Profiles used are
SampleSingleMSPSolo & SampleSingleMSPChannel.  Modify the
orderer profile in configtxyaml and change address from
127.0.0.1:7050 -> orderer:7050 before reading in the file
to configtxgen.  This allows the peer to
resolve with the ordering service when the docker compose
is spun up.

Change CORE_PEER_ADDRESS from 7051 to 7052 when starting
the chaincode in dev mode.

Change-Id: I0dbd3f81553c85943a4219123f086e69081c4f5a
Signed-off-by: Nick Gaski <ngaski@us.ibm.com>