[FAB-10811] fabric-ca sample is broken on v1.2
The Profile section of configtx.yaml must now be at the end of the file,
so the building of this file dynamically must be updated.
I also updated the version of the fabric CA in bootstrap.sh to 1.2.0.

Change-Id: Ifabccebc901f74569cfa1fe1d824fbcd26c4ffd1
Signed-off-by: Keith Smith <bksmith@us.ibm.com>
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>
(cherry picked from commit 75e293183b8f2ccd608d49c5cdcf628b49ee133f)