[FAB-12879] Update fabcar script for new contracts
This CR updates the startFabric.sh script in the fabcar
sample to deploy the new fabcar contracts written with
the new programming model, in either JavaScript or
TypeScript.

Because both JavaScript and TypeScript use the chaincode
language of "node", we need to separate "source" language
from "runtime" language. Also taking the chance to rename
the option "golang" to "go" to match the directory.

Change-Id: Ib42661223925b1d8df8fd73cf8c1aac9d8a29e59
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>