[FAB-16036] Fix a error of get query in testAPI.sh
This CR add a space in parameter expansion to get a transation id.
testAPI.sh tries to get a transaction id using parameter expansion from
MESSAGE variable and accesses a URL incluse the transaction id. However
the transaction id includes a space at the beginning and the access will
be failed.
By ridding of the space, the access will successes.

Signed-off-by: Nao Nishijima <nao.nishijima.xt@hitachi.com>
Reported-by: YAMAUCHI，KAITO <kaito.yamauchi.kk@hitachi-solutions.com>
Change-Id: I0c52418044deb3b7d8c30cb64a5908e32462bd09