Change Download Location of Fabric Binaries
The current scripts download the fabric binaries and place them
in `/usr/local`. Modifying the host system in CI is generally
bad practice. The `bootstrap.sh` script also downloads the
binaries into the root of the `fabric-samples` repo. It is
good practice for us to do the same in CI. We also have
several samples that make an assumption about the location
of these scripts. In the future we should re-evaluate these
examples.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>