Improve asset-transfer-basic javascript application workflow + bugfix in createAsset chaincode arg sequence
Adding a sample application workflow to demonstrate the chaincodes, similar to the WIP PR for basic go-application (https://github.com/hyperledger/fabric-samples/pull/211).
Plan is for app.js to replace query.js & invoke.js, to simplify and document-in-code the app workflow of this sample.
Bugfix: `createAsset` param sequence updated to match the chaincode-go `CreateAsset`.

Signed-off-by: Sijo Cherian <sijo@ibm.com>