Use Busybox to Remove Files and Folders
We delete folders and files locally on the host system
today. If the user is not a super user the files can't
be deleted as they are created by a root user in the
containers. This change uses busybox to perform the
deletions instead so a root user can remove them.

Signed-off-by: Brett Logan <brett.t.logan@ibm.com>