Fix asset-transfer-basic/application-gateway-go
Error handling example used a transaction invocation with the wrong number of arguments and so did not demonstrate the same error condition as other language samples. The error handling code was also not as comprehensive as it should be.

Also changed the console output throughout to match the content and format of the Node and Java samples.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>