Remove heritage chaincode samples
The samples repository has been refactored to focus on an asset transfer series of samples.
The goal of the samples repository is to align with the fabric documentation tutorials
with minimal other distractions.
The heritage samples were left in place for a period of time until the asset transfer series stabilized.
It is now time to purge the heritage samples.

Signed-off-by: David Enyeart <enyeart@us.ibm.com>