Update configtx.yaml
It seems the profile name was incorrectly changed to from "TwoOrgsApplicationGenesis" as the fabric tutorial requires to "ChannelUsingRafts".  This latter profile name creates an error when setting up channels using the test network and tutorials.  This update changes the profile name back to "TwoOrgsApplicationGenesis" as is required for the test network tutorials for which this file was created.

Signed-off-by: Ken P <83926126+schrodingerscode@users.noreply.github.com>