Simplify Java client error-handling example
The latest fabric-gateway client API release (v1.7.0) includes the gRPC error
details in the GatewayExcetion stack trace so it is not necessary to
programmatically access them to demonstrate that they are present.

This change updates the asset-transfer-basic/application-gateway-java
sample to simplify the updateNonExistentAsset example method. It also
updates all samples to use the latest fabric-gateway release.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>