Fix failing ci pipeline with js chaincode
The Javascript asset-transfer-basic chaincode stores AppraisedValue and
Size as string types instead of number types. This leads to an issue
when used with a Go client application where assets are unmarshaled into
an Asset type where AppraisedValue and Size are of uint64 type.

This change makes sure that AppraisedValue and Size are stringified as
numbers.

To prevent the pipeline from failing when the expected error occurs a
sentinel error was created and checked against in the entry point.

Signed-off-by: Stanislav Jakuschevskij <stas@two-giants.com>